### 2024-05-17 #6760 @crvernon
:wave: @TommyJones - Thank you for your submission to JOSS!  Could you please include your paper and supporting files in a branch of the submission repository (https://github.com/TommyJones/tidylda) instead of in a separate repo?  Thanks.
