+++ tidylda-joss/paper.md
@@ -1,0 +1,107 @@
+---
+title: 'tidylda: An R Package for Latent Dirichlet Allocation Using ''tidyverse''
+  Conventions'
+tags:
+- R
+- topic models
+- LDA
+- natural language processing
+- tidy data
+date: "9 May 2024"
+output: pdf_document
+authors:
+- name: Tommy Jones
+  orcid: "0000-0001-6457-2452"
+  equal-contrib: yes
+  affiliation: 1
+bibliography: paper.bib
+affiliations:
+- name: Foundation, USA
+  index: 1
+---
+
+# Summary
+_tidylda_ is a package for a topic model, Latent Dirichlet Allocation or LDA [@blei2002lda], that is natively compatible with the _tidyverse_ [@tidyverse]. _tidylda_'s Gibbs sampler is written in C++ for performance and offers several novel features. It also has methods for sampling from the posterior of a trained model, for more traditional Bayesian analyses.
+
+_tidylda_ takes its syntactic cues from an ecosystem of R packages known as _the tidyverse_. The tidyverse's goal is to "facilitate a conversation between a human and computer about data" [@tidyverse]. Packages in---and adjacent to---the tidyverse share a common design philosophy and syntax based on "tidy data" principles [@wickham2014tidy]. Tidy data has each variable in a column, each observation in a row, and each observational unit in a table. Extensions include the _broom_ package [@broom] for "tidying" up outputs from statistical models and the in-development _tidymodels_ ecosystem [@tidymodels] which extends the tidyverse philosophy to statistical modeling and machine learning workflows. 
+
+Silge et al. articulated a "tidy data" framework for text analyses---the _tidytext_ package [@tidytextjoss]. Their approach has "one row per document per token". The _tidytext_ package provides functionality to tokenize a corpus, transform it into this "tidy" format, and manipulate it in various ways, including preparing data for input into some of R's many topic modeling packages. The _tidytext_ package also provides tidying functions in the style of _broom_ to harmonize outputs from some of R's topic modeling packages. _tidylda_ manages inputs and outputs in the flavor of _tidytext_ but in one self contained package.
+
+# Statement of Need
+Packages that implement topic models in R and other languages are plentiful. Why do we need another? _tidylda_'s native compatibility with the _tidyverse_ makes it significantly more user friendly than other topic modeling packages. It also enables more traditional Bayesian analyses such as the ability to set more flexible priors, burn-in iterations, averaging over segments of the Gibbs sample chain, and sampling from the posterior that other packages lack. Finally, _tidylda_ implements a transfer learning algorithm developed in [@jones2023latent], unavailable in any other package and described in more detail in the following section.
+
+# Latent Dirichlet Allocation and Notation
+LDA is a Bayesian latent variable model for text [@blei2002lda]. It decomposes a data set of word counts, $\boldsymbol{X}$, whose row/column entries, $d,v$, represent the number of times word $v$ is found in document $d$, into two matrices: $\boldsymbol\Theta$ and $\boldsymbol{B}$. The former gives a distribution of (latent) topics over documents and the latter gives a distribution of words over topics. Formally, LDA is
+
+\begin{align}
+  z_{d_{n}}|\boldsymbol\theta_d &\sim 
+  \text{Categorical}(\boldsymbol\theta_d)\\
+  w_{d_{n}}|z_{k},\boldsymbol\beta_k^{(t)} &\sim
[…]
