+++ tidylda-joss/paper.bib
@@ -1,0 +1,166 @@
+@misc{text2vec, 
+author = {Selivanov, Dmitriy and Bickel, Manuel and Wang, Qing}, 
+title = {{text2vec}}, 
+url = {https://CRAN.R-project.org/package=text2vec}, 
+publisher = {CRAN}, 
+copyright = {GPL-3}, 
+year = {2020}
+}
+@article{tidyverse, 
+title = {{Welcome to the tidyverse}}, 
+author = {Wickham, Hadley and Averick, Mara and Bryan, Jennifer and Chang, Winston and McGowan, Lucy D'Agostino and François, Romain and Grolemund, Garrett and Hayes, Alex and Henry, Lionel and Hester, Jim and Kuhn, Max and Pedersen, Thomas Lin and Miller, Evan and Bache, Stephan
+ Milton and Müller, Kirill and Ooms, Jeroen and Robinson, David and Seidel, Dana Paige and Spinu, Vitalie and Takahashi, Kohske and Vaughan, Davis and Wilke, Claus and Woo, Kara and Yutani, Hiroaki}, 
+journal = {Journal of Open Source Software}, 
+doi = {10.21105/joss.01686}, 
+pages = {1686}, 
+number = {43}, 
+volume = {4}, 
+year = {2019}
+}
+@article{roberts2019stm, 
+title = {{stm : An R Package for Structural Topic Models}}, 
+author = {Roberts, Margaret E and Stewart, Brandon M and Tingley, Dustin}, 
+journal = {Journal of Statistical Software}, 
+doi = {10.18637/jss.v091.i02}, 
+number = {2}, 
+volume = {91}, 
+year = {2019}
+}
+@article{tidytextjss, 
+title = {{Text Mining with R: A Tidy Approach}}, 
+author = {Fay, Colin}, 
+journal = {Journal of Statistical Software}, 
+doi = {doi: 10.18637/jss.v083.b01}, 
+number = {Book Review 1}, 
+volume = {83}, 
+year = {2018}
+}
+@article{quanteda, 
+title = {{quanteda: An R package for the quantitative analysis of textual data}}, 
+author = {Benoit, Kenneth and Watanabe, Kohei and Wang, Haiyan and Nulty, Paul and Obeng, Adam and Müller, Stefan and Matsuo, Akitaka}, 
+journal = {Journal of Open Source Software}, 
+doi = {10.21105/joss.00774}, 
+pages = {774}, 
+number = {30}, 
+volume = {3}, 
+year = {2018}
+}
+@article{tidytextjoss, 
+title = {{tidytext: Text Mining and Analysis Using Tidy Data Principles in R}}, 
+author = {Silge, Julia and Robinson, David}, 
+journal = {The Journal of Open Source Software}, 
+doi = {10.21105/joss.00037}, 
+pages = {37}, 
+number = {3}, 
+volume = {1}, 
+year = {2016}
+}
+@misc{textminer, 
+author = {Jones, Tommy}, 
+editora = {William Doane}, 
+title = {{textmineR: Functions for Text Mining and Topic Modeling}}, 
+url = {https://CRAN.R-project.org/package=textmineR}, 
+year = {2015}
+}
+@misc{chang2015lda, 
+author = {Chang, Jonathan}, 
+title = {{lda}}, 
+url = {https://CRAN.R-project.org/package=lda}, 
+copyright = {LGPL-2, LGPL-3}, 
+year = {2015}
+}
+@article{broom, 
+title = {{broom: An R Package for Converting Statistical Analysis Objects Into Tidy Data Frames}}, 
+author = {Robinson, David}, 
+journal = {arXiv}, 
+eprint = {1412.3565}, 
[…]
