+++ tidylda-joss/README.md
@@ -1,0 +1,1 @@
+Repository for JOSS submission of [`tidylda`](https://github.com/tommyjones/tidylda)
