### 2024-05-18 #6760 @crvernon
@TommyJones I think you may need to add brackets around your dates on this ref to get it to compile:
```
@online{tidymodels,
  author = {Max Khun and Hadley Wickham},
  title = {Tidymodels},
  year = 2018,
  urldate = 2021-01-01,
  url = {https://www.tidymodels.org/}
}
```

Per the error...
> Checking the BibTeX entries failed with the following error:
>
> ```
> Failed to parse BibTeX on value "-01-01" (NAME) [#<BibTeX::Bibliography data=[19]>, "@", #<BibTeX::Entry >, {:author=>["Max Khun and Hadley Wickham"], :title=>["Tidymodels"], :year=>"2018", :urldate=>"2021"}]
> ```
