+++ tidylda-joss/paper.bib
@@ -161,6 +161,6 @@ year = {2003}
   author = {Max Khun and Hadley Wickham},
   title = {Tidymodels},
   year = 2018,
-  urldate = 2021-01-01,
+  urldate = {2021-01-01},
   url = {https://www.tidymodels.org/}
 }
