### 2024-05-18 #6760 @TommyJones
Do you still need me to suggest reviewers? Both Emil Hvitfeldt (EmilHvitfeldt) and Julia Silge (juliasilge) would be excellent.

Of the people who have pre-voluneered to review papers, both [Harish Varma Siravuri](https://reviewers.joss.theoj.org/reviewers/3361) and [@bagheria](https://reviewers.joss.theoj.org/reviewers/521) are tagged with all three of R, C++, and natural language processing.
