+++ tidylda-joss/paper.bib
@@ -30,7 +30,7 @@ year = {2019}
 title = {{Text Mining with R: A Tidy Approach}}, 
 author = {Fay, Colin}, 
 journal = {Journal of Statistical Software}, 
-doi = {doi: 10.18637/jss.v083.b01}, 
+doi = {10.18637/jss.v083.b01}, 
 number = {Book Review 1}, 
 volume = {83}, 
 year = {2018}
