### 2024-05-24 #6760 @kanishkan91
@maximelenormand Given your publication history in JOSS, would you be interested in reviewing this software paper for JOSS?
