### 2024-05-24 #6760 @maximelenormand
Hi @kanishkan91, I am available to review this paper for JOSS.
