### 2024-05-28 #6800 @editorialbot
[REVIEW]: tidylda: An R Package for Latent Dirichlet Allocation Using ''tidyverse''   Conventions
**Submitting author:** @tommyjones (<a href="http://orcid.org/0000-0001-6457-2452">Tommy Jones</a>)
**Repository:** https://github.com/TommyJones/tidylda
**Branch with paper.md** (empty if default branch): main
**Version:** v0.06.999
