### 2024-06-11 TommyJones/tidylda#72 @maximelenormand
JOSS review: Duplicated license files
Hi Tommy,

My name is Maxime Lenormand. I will review your package over the next few weeks.

You have two license files, I recommend keeping only the MIT License file.

[#6800](https://github.com/openjournals/joss-reviews/issues/6800)
