### 2024-06-12 254f58e
preparing to make changes based on JOSS submission
