### 2024-06-12 881f92e
Remove duplicated license per #72
