### 2024-06-12 c33a921
remove references to experimental lifecycle
