+++ DESCRIPTION
@@ -28,7 +28,7 @@ Description: Implements an algorithm for Latent Dirichlet
     and 'tidymodels', Kuhn et al.<https://tidymodels.github.io/model-implementation-principles/>.
     Fitting is done via collapsed Gibbs sampling.
     Also implements several novel features for LDA such as guided models and
-    transfer learning based on ongoing and, as yet, unpublished research.
+    transfer learning.
 License: MIT + file LICENSE
 URL: https://github.com/TommyJones/tidylda/
 BugReports: https://github.com/TommyJones/tidylda/issues
