### 2024-06-12 fc57512
remove refs to experimental in documentation
