### 2024-06-12 3c1ac7f
Fix bug to License file causing CMD check to fail
