### 2024-06-12 77a04b5
continuing the license CMD check saga
