### 2024-06-12 846adcc
Removing language around MIT license to conform to CMD check standards.
