### 2024-06-12 TommyJones/tidylda#72 @TommyJones
Actually, I'm reopening this.

CRAN has a very esoteric take on license files. Removing the LICENSE file with just my name and date and/or renaming LICENSE.md (which has MIT license language) means that the package no longer passes checks to be hosted on CRAN.

I understand that it's confusing to have both LICENSE and LICENSE.md files. So, I'm leaving this open until I can satisfy both requirements of having one file and passing CMD check for CRAN.
