### 2024-06-12 TommyJones/tidylda#72 @TommyJones
Ok. What works is removing the MIT license file and leaving the file with just a copright statement. The MIT license is a standard anyway, so I don't think removing it from the repo substantially changes understandability of the copyright.

The commit that passes CMD check and has only one license file is https://github.com/TommyJones/tidylda/commit/846adcced612dfcaa26098c077aad9ba53ddc5ff

@maximelenormand is this solution satisfactory?
