### 2024-06-12 TommyJones/tidylda#72 @maximelenormand
The problem now is that the license can only be identified in the DESCRIPTION file.

This is strange that you cannot pass the CRAN checks with one file. I did not encounter this problem with my packages (example https://github.com/EpiVec/TDLM/) but it is maybe due to the MIT license.

If you cannot solve the problem, you can keep the two license files. It is better to clearly identify one well-known license, even if having two files may be confusing.
