### 2024-06-12 TommyJones/tidylda#72 @TommyJones
Understood. I'll revert to what I had originally. I'll close the issue when that's done.

It looks like [other folks](https://stackoverflow.com/questions/43550479/how-to-satisfy-both-cran-and-github-license-file-naming-requirements) have similar issues.

As I'd originally set it up, I followed Hadely's guidance [here](https://r-pkgs.org/license.html), which was also suggested in the Stack Overflow post, above.

This is less than ideal but it seems to be the conventional wisdom.
