### 2024-06-12 39b751a
putting license back into cran shape using usethis::use_mit_license()
