### 2024-06-17 TommyJones/tidylda#73 @maximelenormand
JOSS review: Changelog, README and Vignettes
I managed to install both the CRAN and GitHub versions. In both cases, I ran the examples without any problems.

- The differences between the current CRAN version and the development version are not clear. It would be nice to add a changelog somewhere to list all the changes made between the different (futur) CRAN versions and between the current CRAN version and the development GitHub version. It seems that there is no mention of either the CRAN or development GitHub version nor the way to install the package in the paper.

- The dependencies are listed in the DESCRIPTION file but not in the README.

- The package contains three vignettes that are not presented on the GitHub repository.

This is not mandatory, of course, but a nice way to address these issues could be to build a website for your package using pkgdown (https://pkgdown.r-lib.org/). A tidyverse template is available at https://github.com/tidyverse/tidytemplate.

[#6800](https://github.com/openjournals/joss-reviews/issues/6800)
