### 2024-06-30 TommyJones/tidylda#74 @hassaniazi
JOSS Review: Minor documentation updates
Hi, great job on developing this package. I will post detailed comments on the paper later. For now, I have two minor requests:
1. There are two License files in the repo at the moment. Can you merge those into one. I believe you can modify the MIT license.
2. JoSS wants your documentation of the package to also include some things that you have mentioned in the paper. I think you have mentioned most of it already in the package documentation as well but can you please make sure using the **Documentation** section of this checklist: https://github.com/openjournals/joss-reviews/issues/6800#issuecomment-2183637317. You may port some text from the paper.
