### 2024-06-30 TommyJones/tidylda#75 @hassaniazi
JOSS Review: Paper review comments 
## Review for Latent Dirichlet Allocation Using ‘tidyverse’ Conventions
This [paper](https://github.com/TommyJones/tidylda/blob/main/tidylda-joss/paper.md) describes the implementation of LDA using tidyverse ecosystem to improve user-friendliness and give more control over complex data wrangling. Doing so has enabled new capabilities e.g., non-uniform initializations or setting flexible priors for LDA. Thanks for providing sample data and guidance  for users on how to run to explore the model's functionality. I was able to execute tidy functions (print, glance, augment etc) and reproduce the results and plots from [README.md](https://github.com/TommyJones/tidylda/blob/main/README.md), good job!

#### General comments:
I have two overall comments and then some specific comments:

1. Analytical contributions
It is unclear from the paper if any analytical adjustments to the LDA formulation have been made or if the "novel features" are mostly in the code base to improve the workflow. Initialization from prior Dirichlet or tLDA seem to qualify as analytical modifications that warrant to be highlighted further (e.g., explained in text and supported by an equation) for clarity and visibility.

2. Paper writing
More work needs to go into giving it the paper a more cogent and incisive shape. Part of the summary and some other bits read as a book/dissertation chapter. Transforming the background theory and focusing more on the contributions of this work could be one way to respond. For instance, taking the second para of the summary as an example, I’d recommend adapting the summary to highlight your contributions rather than the summary of tidyverse, which is well established and isn’t the main contribution of this work. Reference to "described in Chapter 5" in line 71 also needs to be removed. To be clear, I am not referring to the very useful _Latent Dirichlet Allocation and Notation_ and _Transfer LDA (tLDA)_  sections and not recommending a full rewrite-up of the paper, just minor but intentional adjustments would suffice.

### Specific comments:

- Line 04 – Affiliation: do you want to specify which foundation?
- Line 18 – just a minor reference formatting comment: Is the reference for Silge et al. missing? If it is referring to Silge and Robinson then it would Silge and Robinson (2016) not Silge et al
- Line 22 – can you add a few words about “broom style” for the readers who might not be familiar with it
- Line 30 – can you add a few examples of other comparable packages?
- Line 69 – A few words introducing the Gibbs sampler and its specific use in tidylda would improve the clarity of the manuscript
- Line 71 – Reference to "described in Chapter 5" needs to be removed.
- Line 74 – Do you mean single discrete values? If not can you clarify the difference between scalar and vector in the context of its usage for tidylda?
- Line 77 – If tLDA is one of the unique contributions of this work then this should be highlighted in the text. Currently, it is not clear.
- Line 80 – Again, it'd be useful to spell out other implementations somewhere in the text.
- Line 107 – Are the arguments in Dirichlet(𝐶𝑣𝑘 + 𝜂𝑘) and Dirichlet(𝐶𝑣𝑘 + 𝜂𝑘) for tLDA meant to be the same?

Review repo: https://github.com/openjournals/joss-reviews/issues/6800
