### 2024-06-30 #6800 @hassaniazi
Hi @kanishkan91, I've finished my review.

- Repo comments: https://github.com/TommyJones/tidylda/issues/74
- Paper comments: https://github.com/TommyJones/tidylda/issues/75
