### 2024-07-03 TommyJones/tidylda#74 @TommyJones
Hi. Thanks for reviewing.

RE the first issue, we discussed this as part of #72. The TLDR is that the current license structure in the package is [consistent with the recommended approach for R packages](https://r-pkgs.org/license.html). It's not ideal, I know, but it's an artifact of CRAN policies.

For the second issue, I'm seeing that I don't have community guidelines but the rest is there. Does that reflect what you are asking for?
