### 2024-07-03 TommyJones/tidylda#73 @TommyJones
Hi. Thanks again for the review.

* I have a [NEWS.md file](https://github.com/TommyJones/tidylda/blob/main/NEWS.md) that contains updates between versions. I add to the list as I make changes so the current state reflects changes to the development version. Does that meet the need you flagged?
* I'd rather not have dependencies listed in two places. I'm afraid that it would introduce the possibility of human error. (i.e., I will forget to update the README.) If acceptance of the package into JOSS hinges on this, I'll add dependencies to README,  but I'd prefer not to. `tidytext`, for example, [does not list its dependencies in the README](https://github.com/juliasilge/tidytext).
* The vignettes are [here](https://github.com/TommyJones/tidylda/tree/main/vignettes). They're in R markdown format as CRAN does the PDF rendering on their end.

Do the above work for you?
