### 2024-07-03 TommyJones/tidylda#75 @TommyJones
Hello. Thank you again for the review. I'll structure my response/actions based on your comments above. I've included a checklist for me at the bottom to take care of these issues. But I do have a need for more guidance as it relates to 1 & 2, below.

### General Comments

1. I'm confused. Lines 49 - 60 of the paper cover the mathematical modification of LDA. Lines 61 - 115 cover the computational aspects. I also cited the dissertation upon which the package is based (Jones, 2023). So can I ask for specific recommendations on what needs clarification here?
2. I will remove any accidental references to my dissertation (e.g. "chapter 5").
    I'm not sure what you mean by giving the paper "a more cogent and incisive shape." I'm again confused as it seems like I have conflicting guidance here. You say "Transforming the background theory and focusing more on the contributions of this work could be one way to respond" but then also say "I am not referring to the very useful Latent Dirichlet Allocation and Notation and Transfer LDA (tLDA) sections." But implementing tLDA is one of the fundamental intellectual contributions here. (I am working on getting a stand alone tLDA paper published in a more appropriate journal as it's outside of the scope of JOSS.) Forgive me; I am not trying to be difficult. I genuinely don't know what I'm being asked to do.

### Specific Comments
* Line 04 – The company is named Foundation.
* Line 18 – I will update to say Silge and Robinson.
* Line 22 – I will add a few words about “broom style” for the readers who might not be familiar with it
* Line 30 – I will add a paragraph about topic modeling packages in R and OSS in general.
* Line 69 – I will add a few sentences on Gibbs sampling as an introduction.
* Line 71 – I will remove any reference to "described in Chapter 5" or similar.
* Line 74 – I am reluctant to make this change. Maybe it's my math background, but "scalar" has a precise meaning. i.e., a number.
* Line 77 – This seems like it relates to 1 & 2, above. I will wait on your response before addressing this.
* Line 80 – I will add a paragraph about topic modeling packages in R and OSS in general.
* Line 107 – They are nearly the same. There is a subscript on eta in the latter because it refers to tLDA. There is no subscript on eta in the former because it refers to vanilla LDA.

### Checklist
- [ ] Clarify intellectual contribution (awaiting guidance)
- [ ] Line 18 - say "Silge and Robinson"
- [ ] Line 22 - describe the broom package in one sentence
- [ ] Add a paragraph about LDA implementations in R and OSS in general
- [ ] Line 69 - describe Gibbs sampling
- [ ] Remove any references to "chapter 5" or similar on line 71 and throughout the paper
