### 2024-07-03 TommyJones/tidylda#73 @maximelenormand
Hi, regarding the changelog and dependencies, it would be beneficial to add a few sentences to the README to inform users that all changes are listed in the NEWS.md file and that the dependencies can be found in the DESCRIPTION.md file.

Please, add a paragraph describing the different vignettes and the way to compile them.

Do not forget to mention the CRAN and development GitHub versions in the paper and the way to install them.
