### 2024-07-03 TommyJones/tidylda#76 @maximelenormand
JOSS review: Comments on the paper
I have two general comments on the paper.

- The summary should focus on your package and not on the tidy universe. Please use this section to describe the functionality and purpose of your package for a diverse, non-specialist audience.

- The state of need/field section could be developed by listing the main packages that implement topic models in R (and other languages) and explaining in detail why your software is much better than the other options.

[#6800](https://github.com/openjournals/joss-reviews/issues/6800)
