### 2024-07-03 b30c5b4
update references based on JOSS comments
