+++ tidylda-joss/paper.md
@@ -23,9 +23,23 @@ affiliations:
 # Summary
 _tidylda_ is a package for a topic model, Latent Dirichlet Allocation or LDA [@blei2002lda], that is natively compatible with the _tidyverse_ [@tidyverse]. _tidylda_'s Gibbs sampler is written in C++ for performance and offers several novel features, such as transfer learning for LDA using the tLDA model. It also has methods for sampling from the posterior of a trained model, for more traditional Bayesian analyses.
 
+## The "tidyverse" family of packages for R
 _tidylda_ takes its syntactic cues from an ecosystem of R packages known as _the tidyverse_. The tidyverse's goal is to "facilitate a conversation between a human and computer about data" [@tidyverse]. Packages in---and adjacent to---the tidyverse share a common design philosophy and syntax based on "tidy data" principles [@wickham2014tidy]. Tidy data has each variable in a column, each observation in a row, and each observational unit in a table. Extensions include the _broom_ package [@broom] for "tidying" up outputs from statistical models and the in-development _tidymodels_ ecosystem [@tidymodels] which extends the tidyverse philosophy to statistical modeling and machine learning workflows. 
 
 Silge and Robinson articulated a "tidy data" framework for text analyses---the _tidytext_ package [@tidytextjoss]. Their approach has "one row per document per token". The _tidytext_ package provides functionality to tokenize a corpus, transform it into this "tidy" format, and manipulate it in various ways, including preparing data for input into some of R's many topic modeling packages. The _tidytext_ package also provides tidying functions in the style of the _broom_ package, which harmonizes outputs from some of R's topic modeling packages into more usable formats. _tidylda_ manages inputs and outputs in the flavor of _tidytext_ but in one self contained package.
+
+## Topic modeling software in R
+R has many packages for topic modeling; none are natively "tidy" though some have wrapper functions available in _tidytext_ that produce tidy outputs. In almost all cases these models support only scalar, or "symmetric", priors for topics over documents. 
+
+The _textmineR_ package [@textminer] is _tidylda_'s predecessor, supporting vector, or "asymmetric"," priors. It supports fitting several topic models, not just LDA. But _textmineR_ does not support transfer learning nor is it consistent with the _tidyverse_ principles. 
+
+The _topicmodels_ package [@topicmodelspackage] supports fitting models for LDA and correlated topic models [@blei2007ctm] with both a collapsed Gibbs sampler and variational expectation maximization (VEM). When using VEM, $\boldsymbol\alpha$ may be treated as a free parameter and estimated during fitting. It only allows users to set symmetric priors. It is designed to be interoperable with the _tm_ package [@tmjss], the oldest framework for text analysis in R. _tidytext_ provides "tidier" functions to make the _topicmodels_ package interoperable with other frameworks, such as _quanteda_ [@quanteda], _text2vec_ [@text2vec], and more. 
+
+The _lda_ package [@chang2015lda] provides a collapsed Gibbs sampler for LDA, and other less well-known models. Its Gibbs sampler is one of the fastest. It allows users to set only symmetric priors. Its syntax is esoteric and it requires text documents as input, but does not offer much flexibility in the way of pre-processing. It is generally not interoperable with other packages without significant programming on the part of its users. 
+
+The _text2vec_ package [@text2vec] is a framework for very fast text pre-processing and modeling. _text2vec_ implements LDA using the WarpLDA algorithm [@chen2015warplda], but it only allows symmetric priors. _text2vec_ also offers other models related to distributional semantics. Its syntax is also esoteric using R's _R6_ objects that reach back to actively running C++ code for performance reasons. One of _text2vec_'s novel features is that it implements many different coherence calculations; most packages implement only one or none.
+
[…]
