### 2024-07-03 cac8d20
add references for new topic model section
