+++ tidylda-joss/paper.bib
@@ -164,3 +164,93 @@ year = {2003}
   urldate = {2021-01-01},
   url = {https://www.tidymodels.org/}
 }
+
+@misc{text2vec, 
+author = {Selivanov, Dmitriy and Bickel, Manuel and Wang, Qing}, 
+title = {{text2vec}}, 
+url = {https://CRAN.R-project.org/package=text2vec}, 
+publisher = {CRAN}, 
+copyright = {GPL-3}, 
+year = {2020}
+}
+
+@article{roberts2019stm, 
+title = {{stm : An R Package for Structural Topic Models}}, 
+author = {Roberts, Margaret E and Stewart, Brandon M and Tingley, Dustin}, 
+journal = {Journal of Statistical Software}, 
+doi = {10.18637/jss.v091.i02}, 
+number = {2}, 
+volume = {91}, 
+local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/references/2020-12-25/stmVignette.pdf}, 
+year = {2019}
+
+@article{quanteda, 
+title = {{quanteda: An R package for the quantitative analysis of textual data}}, 
+author = {Benoit, Kenneth and Watanabe, Kohei and Wang, Haiyan and Nulty, Paul and Obeng, Adam and Müller, Stefan and Matsuo, Akitaka}, 
+journal = {Journal of Open Source Software}, 
+doi = {10.21105/joss.00774}, 
+pages = {774}, 
+number = {30}, 
+volume = {3}, 
+local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/references/2021-01-01/10.21105.joss.00774.pdf}, 
+year = {2018}
+}
+
+@misc{textminer, 
+author = {Jones, Tommy}, 
+editora = {William Doane}, 
+title = {{textmineR: Functions for Text Mining and Topic Modeling}}, 
+url = {https://CRAN.R-project.org/package=textmineR}, 
+year = {2015}
+}
+@misc{chang2015lda, 
+author = {Chang, Jonathan}, 
+title = {{lda}}, 
+url = {https://CRAN.R-project.org/package=lda}, 
+copyright = {LGPL-2, LGPL-3}, 
+year = {2015}
+}
+
+@article{topicmodelspackage, 
+title = {{topicmodels:  An R Package for Fitting Topic Models}}, 
+author = {Grün, Bettina and Hornik, Kurt}, 
+journal = {Journal of Statistical Software}, 
+number = {13}, 
+volume = {40}, 
+local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/references/2021-01-01/topicmodels.pdf}, 
+year = {2011}
+}
+
+@article{tmjss, 
+title = {{Text Mining Infrastructure in R}}, 
+author = {Feinerer, Ingo and Hornik, Kurt and Meyer, David}, 
+journal = {Journal of Statistical Software}, 
+number = {5}, 
+volume = {25}, 
+local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/references/2021-01-01/textmining.pdf}, 
+year = {2008}
+}
+
+@article{blei2007ctm, 
+title = {{A correlated topic model of Science}}, 
+author = {Blei, David M. and Lafferty, John D.}, 
+journal = {The Annals of Applied Statistics}, 
+issn = {1932-6157}, 
+doi = {10.1214/07-aoas114}, 
[…]
