### 2024-07-03 fce7686
Add sentence about Gibbs sampling
