+++ tidylda-joss/paper.bib
@@ -183,6 +183,7 @@ number = {2},
 volume = {91}, 
 local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/references/2020-12-25/stmVignette.pdf}, 
 year = {2019}
+}
 
 @article{quanteda, 
 title = {{quanteda: An R package for the quantitative analysis of textual data}}, 
