+++ tidylda-joss/paper.bib
@@ -255,3 +255,14 @@ abstract = {{Developing efficient and scalable algorithms for Latent Dirichlet A
 local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/references/download%202020-06-22/1510.08628.pdf}, 
 year = {2015}
 }
+
+@article{roberts2019stm, 
+title = {{stm : An R Package for Structural Topic Models}}, 
+author = {Roberts, Margaret E and Stewart, Brandon M and Tingley, Dustin}, 
+journal = {Journal of Statistical Software}, 
+doi = {10.18637/jss.v091.i02}, 
+number = {2}, 
+volume = {91}, 
+local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/references/2020-12-25/stmVignette.pdf}, 
+year = {2019}
+}
