+++ tidylda-joss/paper.bib
@@ -174,15 +174,11 @@ copyright = {GPL-3},
 year = {2020}
 }
 
-@article{roberts2019stm, 
-title = {{stm : An R Package for Structural Topic Models}}, 
-author = {Roberts, Margaret E and Stewart, Brandon M and Tingley, Dustin}, 
-journal = {Journal of Statistical Software}, 
-doi = {10.18637/jss.v091.i02}, 
-number = {2}, 
-volume = {91}, 
-local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/references/2020-12-25/stmVignette.pdf}, 
-year = {2019}
+@inproceedings{roberts2013stm, 
+author = {Roberts, Margaret E. and Stewart, Brandon M. and Tingley, Dustin and Airoldi, Edoardo M.}, 
+title = {{The Structural Topic Model and Applied Social Science}}, 
+local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/references/download%202020-06-23/stmnips2013.pdf}, 
+year = {2013}
 }
 
 @article{quanteda, 
