### 2024-07-03 TommyJones/tidylda#75 @TommyJones
Issues except for clarifying contribution are taken care of over several commits ending with d04f62177528b165910b01fe4ad8535b92857e7c
