### 2024-07-03 TommyJones/tidylda#73 @TommyJones
Ok. I'll get these taken care of shortly.

- [ ] Reference NEWS.md in README
- [ ] Reference DESCRIPTION in README
- [ ] Add paragraph to README describing vignettes and how to compile them
- [ ] Add paragraph to the JOSS paper describing vignettes
- [ ] Mention CRAN and development versions and how to install them in the JOSS paper
