+++ README.md
@@ -28,10 +28,6 @@ In addition this implementation of LDA allows you to:
 - apply LDA in a transfer-learning paradigm, updating a model’s
   parameters with additional data (or additional iterations)
 
-Note that the seeding of topics and transfer learning are
-**experimental** for now. They are almost-surely useful but their
-behaviors have not been optimized or well-studied. Caveat emptor!
-
 ## Installation
 
 You can install the development version from
