### 2024-07-03 2653884
describe installation from CRAN
