+++ README.md
@@ -30,6 +30,12 @@ In addition this implementation of LDA allows you to:
 
 ## Installation
 
+You can install the latest CRAN release with:
+
+``` r
+install("tidylda")
+```
+
 You can install the development version from
 [GitHub](https://github.com/) with:
 
