### 2024-07-03 e290938
add reference to NEWS.md
