### 2024-07-03 bbd4312
reference dependencies in DESCRIPTION in README
