+++ README.Rmd
@@ -52,7 +52,7 @@ install.packages("remotes")
 remotes::install_github("tommyjones/tidylda")
 ```
 
-
+For a list of dependencies see the DESCRIPTION file.
 
 # Getting started
 
