### 2024-07-03 d4b0965
Reference vignettes in README
