+++ README.Rmd
@@ -241,6 +241,8 @@ print(lda2)
 print(lda)
 ```
 
+There are several vignettes available in [/vignettes](https://github.com/TommyJones/tidylda/tree/main/vignettes). They can be compiled using `knitr` or you can view [PDF versions on CRAN](https://CRAN.R-project.org/package=tidylda).
+
 See NEWS.md for a changelog, including changes from the CRAN release to the development version on Github.
 
 See the "Issues" tab on GitHub to see planned features as well as bug fixes.
