### 2024-07-03 6e8e2f5
move tidyverse and LDA packages to statement of need section per JOSS comments
