### 2024-07-03 8ef6b49
remove superfluous section in intro vignette
