### 2024-07-03 c4eb725
Mention development & CRAN version and vignettes in paper
