### 2024-07-03 TommyJones/tidylda#73 @TommyJones
All of the above are taken care of as of this commit: c4eb725b095f3fd58be6bd3657c9825cc72a98b1
