### 2024-07-03 1db7a97
Add contribution guidelines and code of conduct
