### 2024-07-03 TommyJones/tidylda#74 @TommyJones
Added contribution guideline and code of conduct here: 1db7a975957d390ba2f3ce96c5f2882564aace26

I'll close this issue upon confirmation that this was the only documentation issue or when I know what other issues need addressing.
