### 2024-07-03 TommyJones/tidylda#76 @TommyJones
I believe I addressed these issues with this commit 6e8e2f53a43d360fa0324e445b09e43bbd6ef8a8

Can you confirm that's sufficient (or not)? I'll leave this issue open until then.
