### 2024-07-03 TommyJones/tidylda#76 @TommyJones
Awesome. Thank you so much for your time and effort in reviewing and making both the paper and the tidylda package better.
