### 2024-07-15 TommyJones/tidylda#74 @hassaniazi
Thanks, looking at the README again, I agree you have all the elements, albeit some are more readily accessible than others, but you can consider this resolved. . Thanks for checking it again.
