### 2024-07-15 TommyJones/tidylda#75 @hassaniazi
Thanks for resolving most of my comments.

1. I wasn't sure if the modifications are novel to this work or have been presented before. In my understanding, these are novel to this. I believe including a phrase like the following should be sufficient: "In this work, we introduce transfer LDA (tLDA), which modifies LDA in the following way:"

2. This comment of mine has also been addressed with your changes through out the manuscript and by creating a new section on topic modeling in R. I was mainly looking for a clear distinction between background and your contributions, which the new paper separates well.

Congratulations on this work.
