### 2024-07-16 TommyJones/tidylda#75 @TommyJones
Awesome. Thank you for taking the time and effort to go through the paper and code. Onward!

Sent with Shortwave <https://www.shortwave.com?utm_medium=email&utm_content=signature&utm_source=am9uZXMudGhvcy53QGdtYWlsLmNvbQ==>
