### 2024-07-20 #6800 @TommyJones
Sorry for the delay. I’ve been teaching workshops all day. Will take care of this ASAP. 🙏

Sent with Shortwave <https://www.shortwave.com?utm_medium=email&utm_content=signature&utm_source=am9uZXMudGhvcy53QGdtYWlsLmNvbQ==>
