### 2024-07-24 #6800 @TommyJones
Hi! I've created a [release of the JOSS accepted package](https://github.com/TommyJones/tidylda/releases/tag/JOSS). Now looking into archiving in Zendo. Will update shortly. 👍
