### 2024-07-24 #6800 @TommyJones
I've uploaded the file to Zenodo. The DOI is 10.5281/zenodo.12809519

Are we all set?
