### 2024-07-24 #6800 @danielskatz
at least one of these is a reference that is duplicated in the .bib file - I suspect they all might be
