### 2024-07-24 #6800 @kanishkan91
@TommyJones It seems several entries in this file -https://github.com/TommyJones/tidylda/blob/main/tidylda-joss/paper.bib are duplicated. Would you mind taking a look and correcting these? This seems to obstruct me from accepting this paper
