### 2024-07-24 #6800 @TommyJones
Absolutely. I’ll get right on it.

Sent with Shortwave <https://www.shortwave.com?utm_medium=email&utm_content=signature&utm_source=am9uZXMudGhvcy53QGdtYWlsLmNvbQ==>
