### 2024-07-24 91a2753
remove duplicate bib entries
