+++ tidylda-joss/paper.bib
@@ -1,11 +1,4 @@
-@misc{text2vec, 
-author = {Selivanov, Dmitriy and Bickel, Manuel and Wang, Qing}, 
-title = {{text2vec}}, 
-url = {https://CRAN.R-project.org/package=text2vec}, 
-publisher = {CRAN}, 
-copyright = {GPL-3}, 
-year = {2020}
-}
+
 @article{tidyverse, 
 title = {{Welcome to the tidyverse}}, 
 author = {Wickham, Hadley and Averick, Mara and Bryan, Jennifer and Chang, Winston and McGowan, Lucy D'Agostino and François, Romain and Grolemund, Garrett and Hayes, Alex and Henry, Lionel and Hester, Jim and Kuhn, Max and Pedersen, Thomas Lin and Miller, Evan and Bache, Stephan
@@ -55,13 +48,7 @@ number = {3},
 volume = {1}, 
 year = {2016}
 }
-@misc{textminer, 
-author = {Jones, Tommy}, 
-editora = {William Doane}, 
-title = {{textmineR: Functions for Text Mining and Topic Modeling}}, 
-url = {https://CRAN.R-project.org/package=textmineR}, 
-year = {2015}
-}
+
 @misc{chang2015lda, 
 author = {Chang, Jonathan}, 
 title = {{lda}}, 
@@ -84,14 +71,7 @@ url = {http://www.R-project.org/},
 address = {Vienna, Austria}, 
 year = {2013}
 }
-@article{topicmodelspackage, 
-title = {{topicmodels:  An R Package for Fitting Topic Models}}, 
-author = {Grün, Bettina and Hornik, Kurt}, 
-journal = {Journal of Statistical Software}, 
-number = {13}, 
-volume = {40}, 
-year = {2011}
-}
+
 @inproceedings{gensim, 
 author = {rek, Radim Reh u and Sojka, Petr}, 
 title = {{Software Framework for Topic Modelling with Large Corpora}}, 
@@ -104,14 +84,7 @@ language = {English},
 note = {\textbackslashurlhttp://is.muni.cz/publication/884893/en}, 
 year = {2010}
 }
-@article{tmjss, 
-title = {{Text Mining Infrastructure in R}}, 
-author = {Feinerer, Ingo and Hornik, Kurt and Meyer, David}, 
-journal = {Journal of Statistical Software}, 
-number = {5}, 
-volume = {25}, 
-year = {2008}
-}
+
 @unpublished{mallet, 
 author = {McCallum, Andrew Kachites}, 
 title = {{MALLET: A Machine Learning for Language Toolkit}}, 
@@ -181,18 +154,6 @@ local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/refer
 year = {2013}
 }
 
-@article{quanteda, 
-title = {{quanteda: An R package for the quantitative analysis of textual data}}, 
-author = {Benoit, Kenneth and Watanabe, Kohei and Wang, Haiyan and Nulty, Paul and Obeng, Adam and Müller, Stefan and Matsuo, Akitaka}, 
-journal = {Journal of Open Source Software}, 
-doi = {10.21105/joss.00774}, 
-pages = {774}, 
-number = {30}, 
-volume = {3}, 
-local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/references/2021-01-01/10.21105.joss.00774.pdf}, 
-year = {2018}
-}
-
 @misc{textminer, 
 author = {Jones, Tommy}, 
 editora = {William Doane}, 
@@ -200,13 +161,7 @@ title = {{textmineR: Functions for Text Mining and Topic Modeling}},
[…]
