### 2024-07-24 #6800 @kanishkan91
@TommyJones I have recommended this for acceptance again now. I will be reading through the paper for typos etc. shortly. The AEiC in this submission track will review shortly and if all goes well this will go live soon!

Once again, **Big** thank you to @hassaniazi and @maximelenormand for reviewing! JOSS is volunteer run and relies heavily on researchers such as yourself.
