### 2024-07-25 #6800 @crvernon
:wave: @TommyJones - I just need you to address the following before I can move to accept:

- [ ] The license specified in your Zenodo archive does not match the one used in your repository.  Please edit the metadata in the Zenodo archive to fix this.  A new release is not necessary.
- [ ] Your GitHub releases show a version of v0.0.3 but you specify a v1 release in the archive.  Could you please clarify?

- [ ] You have equal authorship specified though you are the sole author.  Please remove this designation.
- [ ] Is your affiliation correct?  "Foundation, USA"
- [ ] LINE 41:  there seems to be an extra quote mark in this
- [ ] LINE 156:  to maintain consistency with the rest of your paper headings, the "d" in "details" should be capitalized.
- [ ] Ensure that capitalization is preserved in your references.  You can do this using curly brackets around what you wish to maintain the formatting of in your bib file.  For example, the "D" and "A" should be capitalized in the Jones 2023 reference in "...dirichlet allocation...".  And "r" should be capitalized in the Khun 2019 reference.  Please sweep through these to confirm this for all.
- [ ] Check for any missing DOI that need to be included in your references as reported in an above comment in this thread.  Some may not be available.

I will then move forward with accepting this paper for publication.
