### 2024-07-25 fdedd15
updates based on JOSS comments
