+++ tidylda-joss/paper.bib
@@ -54,11 +54,13 @@ author = {Chang, Jonathan},
 title = {{lda}}, 
 url = {https://CRAN.R-project.org/package=lda}, 
 copyright = {LGPL-2, LGPL-3}, 
+doi = {10.32614/CRAN.package.lda},
 year = {2015}
 }
 @article{broom, 
 title = {{broom: An R Package for Converting Statistical Analysis Objects Into Tidy Data Frames}}, 
 author = {Robinson, David}, 
+doi = {10.32614/CRAN.package.broom},
 journal = {arXiv}, 
 eprint = {1412.3565}, 
 abstract = {{The concept of "tidy data" offers a powerful framework for structuring data to ease manipulation, modeling and visualization. However, most R functions, both those built-in and those found in third-party packages, produce output that is not tidy, and that is therefore difficult to reshape, recombine, and otherwise manipulate. Here I introduce the broom package, which turns the output of model objects into tidy data frames that are suited to further analysis, manipulation, and visualization with input-tidy tools. Broom defines the "tidy", "augment" and "glance" generics, which arrange a model into three levels of tidy output respectively: the component level, the observation level, and the model level. I provide examples to demonstrate how these generics work with tidy tools to allow analysis and modeling of data that is divided into subsets, to recombine results from bootstrap replicates, and to perform simulations that investigate the effect of varying input parameters.}}, 
@@ -95,6 +97,7 @@ year = {2002}
 title = {{The Comprehensive R Archive Network}}, 
 author = {Team, The CRAN}, 
 url = {https://cran.r-project.org/}, 
+doi = {10.1002/wics.1212},
 urldate = {2020-01-01}
 }
 @article{blei2002lda, 
@@ -106,7 +109,7 @@ year = {2003}
 }
 
 @phdthesis{jones2023latent,
-  title={Latent Dirichlet Allocation for Natural Language Statistics},
+  title={{Latent Dirichlet Allocation for Natural Language Statistics}},
   author={Tommy Jones},
   year={2023},
   school={George Mason University}
@@ -114,14 +117,14 @@ year = {2003}
 
 @online{tidymodelsbook,
   author = {Khun, Max},
-  title = {Conventions for R Modeling Packages},
+  title = {{Conventions for R Modeling Packages}},
   year = {2019},
   url = {https://tidymodels.github.io/model-implementation-principles/index.html},
   urldate = {2022-04-12}
 }
 
 @article{wickham2014tidy,
-  title={Tidy data},
+  title={{Tidy data}},
   author={Wickham, Hadley and others},
   journal={Journal of Statistical Software},
   volume={59},
@@ -133,6 +136,7 @@ year = {2003}
 @online{tidymodels,
   author = {Max Khun and Hadley Wickham},
   title = {Tidymodels},
+  doi = {10.32614/CRAN.package.tidymodels},
   year = 2018,
   urldate = {2021-01-01},
   url = {https://www.tidymodels.org/}
@@ -144,12 +148,14 @@ title = {{text2vec}},
 url = {https://CRAN.R-project.org/package=text2vec}, 
 publisher = {CRAN}, 
 copyright = {GPL-3}, 
+doi = {10.32614/CRAN.package.text2vec},
 year = {2020}
 }
 
 @inproceedings{roberts2013stm, 
 author = {Roberts, Margaret E. and Stewart, Brandon M. and Tingley, Dustin and Airoldi, Edoardo M.}, 
 title = {{The Structural Topic Model and Applied Social Science}}, 
+doi = {10.32614/CRAN.package.stm},
 local-url = {file://localhost/Users/tommy/Google%20Drive/gmu/_dissertation/references/download%202020-06-23/stmnips2013.pdf}, 
 year = {2013}
 }
[…]
