### 2024-07-25 #6800 @TommyJones
@crvernon, I've addressed these issues as described below. (Excuse the list. GitHub isn't letting me check boxes in others' comments.)

* Updated the Zenodo archive license
* I botched some metadata on my Github release showing an old version as the latest. I've updated it to reflect JOSS as the latest release. Zenodo also defaulted to v1 and I didn't catch it. Updated to the *real* semantic version, v0.06.999, that is the JOSS accepted version.
* Removed equal contribution flag.
* My affiliation is correct.
* Removed the quote.
* Updated capitalization in the title.
* I've used brackets to preserve capitalization in references.
* I've added every DOI that I can find. Many are still unavailable.

I've pushed the latest version to the repo. Should be good to check again.

Thank you!
