### 2024-07-25 #6800 @TommyJones
You all are amazing! Thank you!

Sent with Shortwave <https://www.shortwave.com?utm_medium=email&utm_content=signature&utm_source=am9uZXMudGhvcy53QGdtYWlsLmNvbQ==>
